:white_check_mark: 添加反序列器单元测试函数
