+++ tools/CombatGenerator/tests/data/json_to_combat.json
@@ -1,0 +1,91 @@
+{
+  "UniversalMirageCombatElysianPreheat": {
+    "recognition": "TemplateMatch",
+    "template": "Combat/StopCombat.png",
+    "roi": [
+      0,
+      0,
+      136,
+      140
+    ],
+    "action": "Custom",
+    "pre_delay": 1000,
+    "post_delay": 1000,
+    "next": [
+      "UniversalMirageCombatElysianFinish",
+      "UniversalMirageCombatElysian_001"
+    ],
+    "custom_action": "BasicATK"
+  },
+  "UniversalMirageCombatElysian_001": {
+    "recognition": "TemplateMatch",
+    "template": "Combat/StopCombat.png",
+    "roi": [
+      0,
+      0,
+      136,
+      140
+    ],
+    "action": "Custom",
+    "pre_delay": 0,
+    "post_delay": 30,
+    "next": [
+      "UniversalMirageCombatElysianFinish",
+      "UniversalMirageCombatElysian_002"
+    ],
+    "custom_action": "BasicATK"
+  },
+  "UniversalMirageCombatElysian_002": {
+    "recognition": "TemplateMatch",
+    "template": "Combat/StopCombat.png",
+    "roi": [
+      0,
+      0,
+      136,
+      140
+    ],
+    "action": "Custom",
+    "pre_delay": 0,
+    "post_delay": 100,
+    "next": [
+      "UniversalMirageCombatElysianFinish",
+      "UniversalMirageCombatElysian_003"
+    ],
+    "custom_action": "ExtraSkill"
+  },
+  "UniversalMirageCombatElysian_003": {
+    "recognition": "TemplateMatch",
+    "template": "Combat/StopCombat.png",
+    "roi": [
+      0,
+      0,
+      136,
+      140
+    ],
+    "action": "Custom",
+    "pre_delay": 30,
+    "post_delay": 0,
+    "next": [
+      "UniversalMirageCombatElysianFinish",
+      "UniversalMirageCombatElysian_004"
+    ],
+    "custom_action": "ExtraSkill"
+  },
+  "UniversalMirageCombatElysian_004": {
+    "recognition": "TemplateMatch",
+    "template": "Combat/StopCombat.png",
+    "roi": [
+      0,
+      0,
+      136,
+      140
+    ],
+    "action": "Custom",
+    "pre_delay": 0,
+    "post_delay": 50,
+    "next": [
+      "UniversalMirageCombatFinish"
+    ],
+    "custom_action": "Forward"
+  }
+}
